modified:   Dockerfile 	modified:   build.sh 	modified:   docker-compose.yaml 	modified:   sdk/production/install_production.sh 	modified:   src/go.mod

+++ build.sh
@@ -5,34 +5,44 @@ rm -rf /app/bin
 
 mkdir /app/bin
 
-go mod init app.go
-go mod tidy
 
-go get github.com/joho/godotenv v1.5.1
-go get github.com/google/uuid
-go get github.com/lib/pq v1.10.0 
+# go get github.com/joho/godotenv v1.5.1
+# go get github.com/google/uuid
+# go get github.com/lib/pq v1.10.0 
 
 #  Install KalkanCrypt
-go get github.com/gokalkan/gokalkan
-source /app/sdk/production/install_production.sh
-cp /app/sdk/libkalkancryptwr-64.so.2.0.3 /usr/lib
-mv /usr/lib/libkalkancryptwr-64.so.2.0.3 /usr/lib/libkalkancryptwr-64.so
+# go get github.com/gokalkan/gokalkan v1.3.1
+# source /app/sdk/production/install_production.sh
+# cp /app/sdk/libkalkancryptwr-64.so.2.0.3 /usr/lib
+# mv /usr/lib/libkalkancryptwr-64.so.2.0.3 /usr/lib/libkalkancryptwr-64.so
+cd src
 
-cd /app/src
+# install go modules
+RUN go get github.com/joho/godotenv
+RUN go get github.com/google/uuid
+RUN go get github.com/lib/pq 
+RUN go get github.com/gokalkan/gokalkan
 
+go mod tidy
+
+# go build -o /app/bin/app app.go
+echo "Build start"
 go build -o /app/bin/app app.go
-
-ls -l /app/bin/app
-
-echo "Build complete"
-
-sleep 5
-
-rm -rf /app/src
-rm -rf /app/build.sh
-rm -rf /app/go.mod
-rm -rf /app/go.sum
-
-echo "Clean up complete"
-
-sleep 5
+if [ $? -eq 0 ]; then
+    # mv /app/src/app /app/bin/app
+    sleep 5
+    if [ -f /app/bin/app ]; then
+        echo "Build complete"
+        sleep 5
+        rm -rf /app/src
+        rm -rf /app/build.sh
+        rm -rf /app/go.mod
+        rm -rf /app/go.sum
+        echo "Clean up complete"
+        sleep 5
+    fi
+else
+    echo "Build failed"
+    sleep 5
+    exit 1
+fi
